// src/redux/actions/types/campaignActionTypes.ts
export enum ECampaignActionTypes {
  LOADING = 'CAMPAIGN/LOADING',
  SET = 'CAMPAIGN/SET',
  UPDATE = 'CAMPAIGN/UPDATE',
  SET_STATUS = 'CAMPAIGN/SET_STATUS',
  ADD_SINGLE_STATUS = 'CAMPAIGN/ADD_SINGLE_STATUS',
  SELECT_CAMPAIGN_GROUP = 'CAMPAIGN/SELECT_CAMPAIGN_GROUP',
  SET_PROPERTIES = 'CAMPAIGN/SET_PROPERTIES',
  ADD_PROPERTIES_SAVE_CONFIG = 'CAMPAIGN/ADD_PROPERTIES_SAVE_CONFIG',
  SET_SELECTED_CAMPAIGN_STATUS = 'CAMPAIGN/SET_SELECTED_CAMPAIGN_STATUS',
  SET_SEARCH = 'CAMPAIGN/SET_SEARCH',
  SET_TOTAL_CAMPAIGNS_COUNT = 'CAMPAIGN/SET_TOTAL_CAMPAIGNS_COUNT',
  SET_CAMPAIGNS_TOGGLE = 'CAMPAIGN/SET_CAMPAIGNS_TOGGLE',
  SET_STATUS_FILTER = 'CAMPAIGN/SET_STATUS_FILTER',
  UPDATE_ROOT = 'CAMPAIGN/UPDATE_ROOT',
}